// src/types/car.ts
export interface CarProps {
	license_plate_number: string
	license_plate_number_score: number
	license_plate_country: string
	license_plate_country_score: number
	car_brand: string
	car_brand_score: number
	car_color: string
	car_color_score: number
	car_type_body: string
	car_type_body_score: number
	url: string
}